+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### DEV
 
+ * Correctly assign the form CSS ID (see #956).
  * Fix the referer management in the back end (see contao/core#6127).
  * Also check for a front end user during header replay (see #1008).
  * Encode the username when opening the front end preview as a member (see contao/core#8762).